+++ CHANGELOG.md
@@ -6,10 +6,12 @@ All notable changes to the "aiXcoder" extension will be documented in this file.
 - Now supports PHP. Works best with "intelliphense" extension.
 
 ### Changed
+- Limit sort results with the same word to the most probable single result.
 
 ### Fixed
 - A bug that sometimes writes the star in the code.
 - Add a space after "for" and "while".
+- Fixed compability with C/C++ Extension insider.
 
 ## [0.1.3] - 2019-06-27
 ### Added